+++ src/ui/control/navigation_control.ts
@@ -94,8 +94,8 @@ export class NavigationControl implements IControl {
 
     _rotateCompassArrow = () => {
         const rotate = this.options.visualizePitch ?
-            `scale(${1 / Math.pow(Math.cos(this._map.transform.pitch * (Math.PI / 180)), 0.5)}) rotateX(${this._map.transform.pitch}deg) rotateZ(${this._map.transform.angle * (180 / Math.PI)}deg)` :
-            `rotate(${this._map.transform.angle * (180 / Math.PI)}deg)`;
+            `scale(${1 / Math.pow(Math.cos(this._map.transform.pitchInRadians), 0.5)}) rotateX(${this._map.transform.pitch}deg) rotateZ(${-this._map.transform.bearing}deg)` :
+            `rotate(${-this._map.transform.bearing}deg)`;
 
         this._compassIcon.style.transform = rotate;
     };